// src/types/api.ts
export interface CoinGeckoAsset {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  total_volume: number;
  price_change_percentage_24h: number;
  platforms?: Record<string, string>;
  categories?: string[];
  chain?: string;
}

export interface ApiConfig {
  urls: {
    [key: string]: string;
  };
  retryAttempts: number;
  staleTime: number;
  refetchInterval: number;
}